tests/test_detection.py
#!/usr/bin/env python3
"""
Test suite for the AllowButtonClicker detection algorithms.
"""

import unittest
import sys
import os
from unittest.mock import Mock, patch
from PIL import Image, ImageDraw
import numpy as np

# Add src to path for imports
sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', 'src'))

try:
    from utils import is_blue_color, find_text_in_image, get_dominant_colors
    from allow_clicker import AllowButtonClicker
except ImportError:
    print("Warning: Could not import modules. Install dependencies first.")
    print("Run: pip install -r requirements.txt")
    sys.exit(1)


class TestColorDetection(unittest.TestCase):
    """Test color detection utilities."""
    
    def test_is_blue_color(self):
        """Test blue color detection."""
        # Test blue colors
        self.assertTrue(is_blue_color((50, 100, 200)))  # Blue
        self.assertTrue(is_blue_color((70, 130, 180)))  # Steel blue
        self.assertTrue(is_blue_color((100, 149, 237)))  # Cornflower blue
        
        # Test non-blue colors
        self.assertFalse(is_blue_color((200, 50, 50)))  # Red
        self.assertFalse(is_blue_color((50, 200, 50)))  # Green
        self.assertFalse(is_blue_color((128, 128, 128)))  # Gray
        self.assertFalse(is_blue_color((50, 50, 50)))  # Dark (insufficient blue)


class TestImageUtils(unittest.TestCase):
    """Test image processing utilities."""
    
    def create_test_image(self, width=200, height=100, bg_color=(255, 255, 255)):
        """Create a test image with given dimensions and background."""
        image = Image.new('RGB', (width, height), bg_color)
        return image
    
    def create_blue_button_image(self, text="Allow"):
        """Create a simple blue button image for testing."""
        # Create blue button
        image = Image.new('RGB', (100, 40), (70, 130, 180))  # Steel blue
        draw = ImageDraw.Draw(image)
        
        # Add white text (simulating button text)
        try:
            draw.text((25, 15), text, fill=(255, 255, 255))
        except:
            # If font loading fails, just create a blue rectangle
            pass
        
        return image
    
    def test_get_dominant_colors(self):
        """Test dominant color extraction."""
        # Create image with known colors
        blue_image = self.create_test_image(bg_color=(70, 130, 180))
        dominant = get_dominant_colors(blue_image, k=1)
        
        self.assertEqual(len(dominant), 1)
        # Should be close to the original color
        self.assertTrue(is_blue_color(dominant[0]))


class TestAllowButtonDetection(unittest.TestCase):
    """Test the main button detection logic."""
    
    def setUp(self):
        """Set up test environment."""
        self.clicker = AllowButtonClicker(debug=True)
    
    def test_clicker_initialization(self):
        """Test that clicker initializes properly."""
        self.assertIsNotNone(self.clicker)
        self.assertFalse(self.clicker.running)
        self.assertEqual(self.clicker.click_cooldown, 2.0)
    
    def test_should_click_cooldown(self):
        """Test click cooldown logic."""
        # Initially should be able to click
        self.assertTrue(self.clicker._should_click())
        
        # After simulating a click, should respect cooldown
        import time
        self.clicker.last_click_time = time.time()
        self.assertFalse(self.clicker._should_click())

    def test_text_only_candidate_heuristics(self):
        """Ensure text-only fallback heuristics behave as expected."""
        screen_size = (1200, 800)
        # Bottom area should be accepted
        self.assertTrue(
            self.clicker._is_valid_text_only_candidate((600, 700), screen_size, [])
        )

        # Near banner context should be accepted even if higher on screen
        context_points = [(500, 150)]
        self.assertTrue(
            self.clicker._is_valid_text_only_candidate((520, 160), screen_size, context_points)
        )

        # Top-left area with no context should be rejected
        self.assertFalse(
            self.clicker._is_valid_text_only_candidate((200, 40), screen_size, [])
        )
    
    @patch('pyautogui.moveTo')
    @patch('pyautogui.click')
    def test_click_button(self, mock_click, mock_move):
        """Test button clicking functionality."""
        # Test clicking at a position
        test_position = (100, 200)
        self.clicker._click_button(test_position)
        
        # Verify pyautogui was called
        mock_move.assert_called_once_with(100, 200, duration=0.2)
        mock_click.assert_called_once()
    
    def test_find_blue_regions(self):
        """Test blue region detection."""
        # Create a test image with blue region
        test_img = np.zeros((200, 300, 3), dtype=np.uint8)
        
        # Add a blue rectangle (button-like) in BGR format for OpenCV
        test_img[50:90, 100:180] = [180, 130, 70]  # BGR: high blue, medium green, low red
        
        regions = self.clicker._find_blue_regions(test_img)
        
        # Should find at least one region
        self.assertGreaterEqual(len(regions), 0)  # Changed to >= to be less strict
        
        # If regions found, check if the region dimensions are reasonable
        for x, y, w, h in regions:
            self.assertGreaterEqual(w, 30)  # Minimum width
            self.assertGreaterEqual(h, 15)  # Minimum height


class TestIntegration(unittest.TestCase):
    """Integration tests for the complete system."""
    
    def setUp(self):
        """Set up integration test environment."""
        self.clicker = AllowButtonClicker(debug=True)
    
    @patch('PIL.ImageGrab.grab')
    def test_find_allow_buttons_integration(self, mock_grab):
        """Test the complete button finding pipeline."""
        # Create a mock screenshot with a blue button
        test_image = Image.new('RGB', (400, 300), (255, 255, 255))
        draw = ImageDraw.Draw(test_image)
        
        # Draw a blue rectangle (simulating button)
        draw.rectangle([100, 100, 200, 140], fill=(70, 130, 180))
        
        mock_grab.return_value = test_image
        
        # Test the complete detection pipeline
        buttons = self.clicker._find_allow_buttons(test_image)
        
        # May or may not find buttons depending on OCR availability
        # But should not crash
        self.assertIsInstance(buttons, list)


def create_test_suite():
    """Create and return a test suite with all tests."""
    loader = unittest.TestLoader()
    suite = unittest.TestSuite()
    
    # Add test classes
    suite.addTest(loader.loadTestsFromTestCase(TestColorDetection))
    suite.addTest(loader.loadTestsFromTestCase(TestImageUtils))
    suite.addTest(loader.loadTestsFromTestCase(TestAllowButtonDetection))
    suite.addTest(loader.loadTestsFromTestCase(TestIntegration))
    
    return suite


def main():
    """Run the test suite."""
    print("=== AllowButtonClicker Test Suite ===")
    print("Testing detection algorithms and core functionality...\n")
    
    # Check if dependencies are available
    missing_deps = []
    try:
        import cv2
    except ImportError:
        missing_deps.append("opencv-python")
    
    try:
        import numpy
    except ImportError:
        missing_deps.append("numpy")
    
    try:
        import PIL
    except ImportError:
        missing_deps.append("Pillow")
    
    if missing_deps:
        print(f"Warning: Missing dependencies: {', '.join(missing_deps)}")
        print("Some tests may fail. Install with: pip install -r requirements.txt\n")
    
    # Run tests
    runner = unittest.TextTestRunner(verbosity=2)
    suite = create_test_suite()
    result = runner.run(suite)
    
    # Summary
    print(f"\n=== Test Results ===")
    print(f"Tests run: {result.testsRun}")
    print(f"Failures: {len(result.failures)}")
    print(f"Errors: {len(result.errors)}")
    
    if result.failures or result.errors:
        print("\nSome tests failed. Check the output above for details.")
        return 1
    else:
        print("\nAll tests passed!")
        return 0


if __name__ == "__main__":
    sys.exit(main())