requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pyautogui>=0.9.50
pytesseract>=0.3.10
pynput>=1.7.6